Add SinglePassOTreeDataAnalyzer, leaving DoublePass intact
